speech-to-text/requirements.txt
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
ffmpy==0.2.2
Flask==1.1.1
fuzzywuzzy==0.17.0
google-api-core==1.14.3
google-auth==1.7.1
google-cloud-core==1.0.3
google-cloud-speech==1.2.0
google-cloud-storage==1.23.0
google-resumable-media==0.5.0
googleapis-common-protos==1.6.0
grpcio==1.25.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
m3u8downloader==0.8.0
MarkupSafe==1.1.1
protobuf==3.10.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
python-Levenshtein==0.12.0
pytz==2019.3
requests==2.21.0
rsa==4.0
six==1.13.0
urllib3==1.24.3
wells==1.4.1
Werkzeug==0.16.0
